+++ apps/sensenet/src/components/dashboard/markdown-widget.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Typography } from '@material-ui/core'
 import ReactMarkdown from 'react-markdown'
 import { MarkdownWidget as MarkdownWidgetModel } from '../../services/PersonalSettings'
-import { useStringReplace } from '../../hooks/use-string-replace'
+import { useStringReplace } from '../../hooks'
 
 export const MarkdownWidget: React.FunctionComponent<MarkdownWidgetModel> = props => {
   const replacedContent = useStringReplace(props.settings.content)
