+++ apps/sensenet/src/hooks/index.ts
@@ -8,6 +8,8 @@ export * from './use-personal-settings'
 export * from './use-repository'
 export * from './use-selection-service'
 export * from './use-session'
+export * from './use-string-replace'
 export * from './use-theme'
+export * from './use-version-info'
 export * from './use-widgets'
 export * from './use-wopi'
